Create platform interface for HostedCluster controller.
This is a first step towards a more sustainable model for the code base to extend to multiple platforms.
It creates an interface to be implemented for each platform to handle the HostedController needs.
It add docs for extend support to other platforms.
As a follow up if this model proves to be satisfactory we can extend the interface to support the NodePool controller and handle infra machines lifecycle. Furthermore, we could at some point consider this to be an adhoc platform controller which communicates with the core one through conditions there fore enabling out of tree platform implementations.
