+++ docs/content/developers/how-to/contributing.md
@@ -1,0 +1,1 @@
+../../../../CONTRIBUTING.md
